requirements.txt
matplotlib==3.3.4
opencv-python==4.5.1.48
tensorflow-gpu==2.4.1
sklearn==0.0